requirements.txt
adal==1.2.7
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
anyio==4.9.0
applicationinsights==0.11.9
argcomplete==3.1.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
asn1crypto==0.24.0
azure-appconfiguration==1.1.1
azure-batch==14.2.0
azure-cli @ file:///azure-cli/src/azure-cli
azure-cli-core @ file:///azure-cli/src/azure-cli-core
azure-cli-telemetry @ file:///azure-cli/src/azure-cli-telemetry
azure-common==1.1.22
azure-core==1.35.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.49
azure-graphrbac==0.60.0
azure-identity==1.23.1
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.0.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0b3
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==30.6.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-containerservice==30.0.0
azure-mgmt-core==1.3.2
azure-mgmt-cosmosdb==9.4.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b1
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.0
azure-mgmt-msi==7.0.0
azure-mgmt-netapp==10.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b14
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.4.0
azure-mgmt-redis==14.3.0
azure-mgmt-resource==23.1.0b2
azure-mgmt-search==9.0.0
azure-mgmt-security==5.0.0
azure-mgmt-servicebus==8.2.0
azure-mgmt-servicefabric==1.0.0
azure-mgmt-servicefabricmanagedclusters==2.0.0b6
azure-mgmt-servicelinker==1.2.0b2
azure-mgmt-signalr==2.0.0b1
azure-mgmt-sql==4.0.0b16
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==21.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.2.0
azure-monitor-opentelemetry==1.6.12
azure-monitor-opentelemetry-exporter==1.0.0b40
azure-monitor-query==1.2.0
azure-multiapi-storage==1.2.0
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.18.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
bcrypt==3.2.0
billiard==4.2.1
celery==5.5.3
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.6.0
dnspython==2.7.0
email_validator==2.2.0
fabric==3.2.2
fastapi==0.116.1
fixedint==0.1.6
googleapis-common-protos==1.70.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.7
importlib_metadata==8.6.1
invoke==2.2.0
isodate==0.6.1
javaproperties==0.5.1
Jinja2==3.1.3
jmespath==0.9.5
jsondiff==2.0.0
knack==0.11.0
kombu==5.5.4
MarkupSafe==2.0.1
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4
oauthlib==3.2.2
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-celery==0.56b0
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-proto==1.35.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.52b1
packaging==23.2
paramiko==3.4.0
pbr==5.3.1
pkginfo==1.8.2
portalocker==2.3.2
postgrest==1.1.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==5.9.5
pycomposefile==0.0.30
pycparser==2.19
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments==2.17.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.1
realtime==2.6.0
redis==5.2.1
requests==2.31.0
requests-oauthlib==1.2.0
scp==0.13.2
semver==2.13.0
six==1.16.0
sniffio==1.3.1
sshtunnel==0.1.5
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
structlog==25.4.0
supabase==2.17.0
supafunc==0.10.1
tabulate==0.8.9
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.18
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.1.7
websocket-client==1.3.1
websockets==15.0.1
wrapt==1.16.0
xmltodict==0.12.0
zipp==3.23.0
